New feature: ungribprocessor
A new key 'ungribprocessor' is available to provide a script (receiving
no arguments) , which will run after ungrib. The script/executable must
modify the namelist.wps if new intermediate files are to be processed by
metgrid.

An ungribprocessor.keyword file is expected to be located at the WPS/
directory. The file must be in the bundle or passed using the
wrf4g_files directory to that path.